+++ rust/stackablectl/CHANGELOG.md
@@ -8,6 +8,11 @@ All notable changes to this project will be documented in this file.
 
 - Add shell completions for Nushell and Elvish ([#337]).
 
+### Fixed
+
+- Sort operator versions by semver version instead of alphabetically ([#336]).
+
+[#336]: https://github.com/stackabletech/stackable-cockpit/pull/336
 [#337]: https://github.com/stackabletech/stackable-cockpit/pull/337
 
 ## [24.11.0] - 2024-11-18